Use different key for CFT for those messages that should be received by all subscribers

+++ srcCpp03/RTIDDSImpl.h
@@ -15,7 +15,6 @@
 #include "security/security_default.h"
 #endif
 #include "rti/config/Logger.hpp"
-#include <dds/dds.hpp>
 #include "PerftestTransport.h"
 
 #define RTIPERFTEST_MAX_PEERS 1024
@@ -71,8 +70,8 @@ class RTIDDSImpl : public IMessaging
     bool         _IsReliable;
     bool         _IsMulticast;
     unsigned int _BatchSize;
-    int          _InstanceCount;
-    int          _InstanceMaxCountReader;
+    unsigned long _InstanceCount;
+    unsigned long _InstanceMaxCountReader;
     int          _InstanceHashBuckets;
     int          _Durability;
     bool         _DirectCommunication;
@@ -90,7 +89,7 @@ class RTIDDSImpl : public IMessaging
     int          _peer_host_count;
     dds::core::StringSeq  _peer_host;
     bool         _useCft;
-    int          _instancesToBeWritten;
+    long         _instancesToBeWritten;
     std::vector<unsigned int> _CFTRange;
 
     PerftestTransport _transport;
